+++ CHANGELOG.rst
@@ -15,6 +15,14 @@ v1.8
 - **v1.8.2:** Fix error handing in :meth:`pycyphal.transport.redundant.RedundantOutputSession.send`;
   see `#222 <https://github.com/OpenCyphal/pycyphal/issues/222>`_.
 
+- **v1.8.3:**
+
+  - ``DiagnosticPublisher``: do not instantiate the publisher if the local node is anonymous.
+  - ``publish_soon()``: Do not log error if closed.
+  - ``Client`` and ``Publisher``: fix edge cases related to ``PortClosedError`` when the interface becomes unavailable.
+  - Fix assertion failure during register value coercion.
+  - SocketCAN: close the media instance automatically on unrecoverable errors like ENODEV, ENXIO, EBADF, EBADFD, etc.
+
 v1.7
 ----